Hotfix init command and increased test coverage for ui/file-info

+++ src/models/file-info.js
@@ -1,6 +1,7 @@
 import ejs from 'ejs';
 import fs from 'fs';
 import { outputFileSync } from 'fs-extra';
+import { fileExists } from '../util/fs';
 
 class FileInfo {
   constructor(args) {
@@ -11,18 +12,50 @@ class FileInfo {
   }
 
   writeFile() {
-    const fileContent = this.renderTemplate();
-    outputFileSync(this.mappedPath, fileContent);
-    this.ui.writeCreate(this.mappedPath);
+    this.ui.writeDebug(`Attempting to write file: ${this.mappedPath}`);
+    if (fileExists(this.mappedPath)) {
+      this.ui.writeError(
+        `Not writing file.  File already exists at: ${this.mappedPath}`
+      );
+    } else {
+      const fileContent = this.renderTemplate();
+      this.ui.writeDebug(`fileContent: ${fileContent}`);
+
+      outputFileSync(this.mappedPath, fileContent);
+      this.ui.writeCreate(this.mappedPath);
+    }
+    return;
   }
 
   renderTemplate() {
+    let rendered;
+    this.ui.writeDebug(`rendering template: ${this.originalPath}`);
     const template = fs.readFileSync(this.originalPath, 'utf8');
-    return ejs.render(template, this.templateVariables);
+
+    try {
+      rendered = ejs.render(template, this.templateVariables);
+    } catch (err) {
+      this.ui.writeDebug('couldnt render');
+      err.message += ' (Error in blueprint template: ' + this.originalPath + ')';
+      this.ui.writeError(`error was: ${err.message}`);
+      throw err;
+    }
+    return rendered;
   }
 
   isFile() {
-    return fs.statSync(this.originalPath).isFile();
+    let fileCheck;
+    try {
+      fileCheck = fs.lstatSync(this.originalPath).isFile();
+    } catch (e) {
+      if (e.code === 'ENOENT') {
+        return false;
+      } else {
+        throw e;
+      }
+    }
+    this.ui.writeDebug(`checking file: ${this.originalPath} - ${fileCheck}`);
+    return fileCheck;
   }
 }
